chore: move lsm_range_test into lsm_test
